Improvement to watching
This adds watching to all files that are copied from the users source
material (we don't watch the cadeau files).  However, this trick doesn't
quite work yet because of a bug in fs-extra which is detailed here:

https://github.com/jprichardson/node-fs-extra/pull/52

But assuming that pull request is accepted, then once that fix is
available via npm then the watching should work fine.  Until then,
we wait.