Switched (rather painlessly) to using swig
Swig is very much like Jinja and includes a lot of the powerful features I really liked
in Jinja and was missing in Handlebars.  This is already paying dividencds because I can
define macros right in my slides.  But it also opens to door for extending from templates
in a really easy way too.

+1