This adds clicker support to Cadeau
In response to issue #17, I've added clicker support.  This really meant adding it
to Flowtime (which I did, and submitted as a pull request to make sure it gets
incorporated upstream).  This commit adds the updated flowtime.js as well as
including some metadata to control clicker support.

I've made clicker support on by default simply because it only impacts behavior
I never used before in Flowtime.

This closes #17.